Move DNS1123 validation logic to shared plugin
