+++ frontend/packages/console-shared/src/utils/grammar.ts
