+++ frontend/packages/console-shared/src/utils/validation/strings.ts
@@ -1,0 +1,6 @@
+export const EMPTY_ERROR = 'cannot be empty';
+
+export const DNS1123_START_END_ERROR = 'has to start/end with alphanumeric characters';
+export const DNS1123_START_ERROR = 'has to start with alphanumeric character';
+export const DNS1123_END_ERROR = 'has to end with alphanumeric character';
+export const DNS1123_TOO_LONG_ERROR = 'cannot have more than 253 characters';
