+++ frontend/packages/kubevirt-plugin/src/components/create-vm-wizard/form/form-field-row.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import * as _ from 'lodash';
+import { ValidationObject } from '@console/shared';
 import { getFieldHelp, getFieldId, getFieldTitle } from '../utils/vm-settings-tab-utils';
 import { iGetFieldValue, isFieldHidden, isFieldRequired } from '../selectors/immutable/vm-settings';
 import { iGet, iGetIn, iGetIsLoaded } from '../../../utils/immutable';
 import { FormRow } from '../../form/form-row';
-import { ValidationObject } from '../../../utils/validations/types';
 import { FormFieldContext } from './form-field-context';
 import { FormFieldType } from './form-field';
 import { FormFieldReviewContext } from './form-field-review-context';
