+++ frontend/packages/kubevirt-plugin/src/components/form/k8s-resource-select-row.tsx
@@ -1,9 +1,8 @@
-import { FirehoseResult } from '@console/internal/components/utils';
-import { K8sKind, K8sResourceKind } from '@console/internal/module/k8s';
 import * as React from 'react';
 import { FormSelect, FormSelectOption } from '@patternfly/react-core';
-import { getName } from '@console/shared/src';
-import { ValidationErrorType, ValidationObject } from '../../utils/validations/types';
+import { FirehoseResult } from '@console/internal/components/utils';
+import { K8sKind, K8sResourceKind } from '@console/internal/module/k8s';
+import { getName, ValidationErrorType, ValidationObject } from '@console/shared';
 import { getLoadedData, getLoadError, isLoaded } from '../../utils';
 import { ignoreCaseSort } from '../../utils/sort';
 import { FormRow } from './form-row';
