+++ frontend/packages/kubevirt-plugin/src/utils/validations/strings.ts
@@ -1,13 +1,7 @@
 export const GENERAL_ERROR_MSG = 'An error occurred. Please try again';
-export const EMPTY_ERROR = 'cannot be empty';
 export const START_WHITESPACE_ERROR = 'cannot start with whitespace characters';
 export const END_WHITESPACE_ERROR = 'cannot end with whitespace characters';
 export const URL_INVALID_ERROR = 'has to be a valid URL';
-
-export const DNS1123_START_END_ERROR = 'has to start/end with alphanumeric characters';
-export const DNS1123_START_ERROR = 'has to start with alphanumeric character';
-export const DNS1123_END_ERROR = 'has to end with alphanumeric character';
-export const DNS1123_TOO_LONG_ERROR = 'cannot have more than 253 characters';
 
 export const VIRTUAL_MACHINE_EXISTS =
   'is already used by another virtual machine in this namespace';
