+++ frontend/packages/kubevirt-plugin/src/utils/validations/vm/vm.ts
@@ -1,11 +1,12 @@
 import { TemplateKind } from '@console/internal/module/k8s';
-import { ValidationErrorType, ValidationObject } from '../types';
 import {
-  getValidationObject,
+  asValidationObject,
+  ValidationErrorType,
+  ValidationObject,
   validateDNS1123SubdomainValue,
-  validateEntityAlreadyExists,
-} from '../common';
+} from '@console/shared';
 import { ProvisionSource } from '../../../constants/vm/provision-source';
+import { validateEntityAlreadyExists } from '../common';
 
 export const validateVmLikeEntityName = (
   value: string,
@@ -31,6 +32,6 @@ export const validateUserTemplateProvisionSource = (
   const provisionSourceDetails = ProvisionSource.getProvisionSourceDetails(userTemplate);
 
   return provisionSourceDetails.error
-    ? getValidationObject(`Could not select Provision Source. ${provisionSourceDetails.error}`)
+    ? asValidationObject(`Could not select Provision Source. ${provisionSourceDetails.error}`)
     : null;
 };
